feat: move quizSets to contract, test js config
