+++ next.config.js
@@ -1,4 +1,10 @@
-export default {
+const basePath = process.env.NEXT_PUBLIC_BASE_PATH
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
     reactStrictMode: true,
-    basePath: process.env.NEXT_PUBLIC_BASE_PATH,
+    basePath: basePath,
 };
+
+
+export default nextConfig
