--- src/shared/constants/tsconfig.next.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": { "outDir": "./dist", "rootDir": "." },
-  "references": []
-}
