+++ src/shared/utils/getRoute.ts
@@ -1,3 +1,4 @@
 export function getRoute(quiz: string, level: string): string {
-  return `${process.env.NEXT_PUBLIC_BASE_PATH}/${quiz}/${level}`;
+  const basePath = process.env.NEXT_PUBLIC_BASE_PATH;
+  return `${basePath}/${quiz}/${level}`;
 }
